Split Wizard into general and compliance request specific components
